% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airheads.R
\name{airheads}
\alias{airheads}
\title{Track QC notes and findings}
\usage{
airheads(qc_data_path = NULL, qc_program_path = NULL)
}
\arguments{
\item{qc_data_path}{path to qc dataset. Auto loads qc_data_path if already in environment.}

\item{qc_program_path}{path to qc code file with comments. Default opens active tab}
}
\description{
Retrieve findings and notes from QC program and generate QC email
FINDING: this is a finding or NOTE: this is a note
}
\examples{
\dontrun{
airheads()
}


}
