% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{broken_pk}
\alias{broken_pk}
\title{Example PK dataset for dispositions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3556 rows and 107 columns.
}
\usage{
broken_pk
}
\description{
Dataset based on \code{dmcognigen} example pk dataset with deletion flags added
and data intentionally manipulated for deletion examples
}
\examples{
head(broken_pk)



}
\keyword{datasets}
