% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Initialize a listing report}
\usage{
build_report(output_path = ".", file_name, ...)
}
\arguments{
\item{output_path}{Directory for listing output}

\item{file_name}{Name of listings file}

\item{...}{Extra arguments to be passed to reporter::options_fixed()}
}
\description{
Initialize a listing report
}
\examples{
del_report <- build_report(file_name = "deletion-listings")



}
