% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{check_models}
\alias{check_models}
\title{Check if modeling in underway}
\usage{
check_models(paths = c("../monolix/", "../nm/"))
}
\arguments{
\item{paths}{to model files. Defaults to \code{paths=c("../monolix/","../nm/")}
Provide the path(s) if you're not working in asmbdat directory}
}
\value{
A data frame with software,output,user,min and max modification times
}
\description{
Sometimes it is useful for programmers to know if and when modeling has occurred
}
\examples{
\dontrun{

# if working in asmbdat and looking just for monolix or nonmem, no arguments needed.
  check_models()

# otherwise provide the path(s)
check_models(path = "../misc/sponsor/sponsor-108/123456/d1pkpd/monolix/")
}
}
