% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_chocula.R
\name{cnt_chocula}
\alias{cnt_chocula}
\title{Compare two counts}
\usage{
cnt_chocula(df1, df2, vars, ...)
}
\arguments{
\item{df1}{first data frame to compare}

\item{df2}{second data frame to compare}

\item{vars}{variables to count}

\item{...}{variables to count}
}
\value{
A data frame with counts and differences.
}
\description{
Compare two counts
}
\examples{
\dontrun{

mtcars2 <- mtcars[mtcars$"cyl"> 4, ]

cnt_chocula(mtcars,mtcars2)

}
}
