% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delimp_biscuit.R
\name{delimp_biscuit}
\alias{delimp_biscuit}
\title{delimp_biscuit}
\usage{
delimp_biscuit(delimp_paths, summary_tbl = TRUE)
}
\arguments{
\item{delimp_paths}{a vector of one or more delimp directories}

\item{summary_tbl}{returns a data.frame of delimp files when TRUE. If FALSE ]
the deletion dataset (i.e. data frame of all delimp files) is returned instead}
}
\value{
a nested data frame showing deletion flags in each directory
}
\description{
delimp_biscuit
}
\examples{
if(basename(getwd()) =="asmbdat" &&
dir.exists("../data/delimp") &&
length(list.files("../data/delimp")>=1 )) {

delimp_biscuit(delimp_paths = "../data/delimp", summary_tbl = TRUE)

}

}
