% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Fake Data: Example Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 320 rows and 4 columns.
}
\usage{
fake_data
}
\description{
modified iris dummy dataset
}
\examples{
\dontrun{
head(fake_data)
}
}
\keyword{datasets}
