% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_contents.R
\name{get_missing_combos}
\alias{get_missing_combos}
\title{Get missing combinations}
\usage{
get_missing_combos(data, ...)
}
\arguments{
\item{data}{a sweet_contents data.frame}

\item{...}{provide a \code{name} variable from \code{sweet_contents} where n_missing is > 0}
}
\value{
A vector of missing combinations and a message that shows selection e.g.
"VISIT" where "AMT" is missing
}
\description{
Quickly build vectors of missing combinations to help with further sleuthing
through missing data
}
\examples{
\dontrun{

    broken_pk \%>\%
    sweet_contents(missing_where = VISIT) \%>\%
    get_missing_combos("AMT")







}
}
