% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_listing_to_report.R
\name{listing_table}
\alias{listing_table}
\title{make a listing table}
\usage{
listing_table(listing, title, ...)
}
\arguments{
\item{listing}{Data to be used in the listing. Subset to variables that best portray why the data were deleted.
Selected variables should be source variables if possible}

\item{title}{Title to describe each unique listing in the report. Usually closely resembles DELFNC}

\item{...}{additional arguments passed to \code{\link[reporter:create_table]{reporter::create_table()}}}
}
\value{
A deletion report object that includes all added listings and their titles of class "table_spec" & "list"
}
\description{
\code{listing_table()}
is used to build listing tables for each deletion data subset
listing tables can then be modified further using \code{\link[reporter:define]{reporter::define()}} or sent to \code{\link[=add_listing_to_report]{add_listing_to_report()}} where
listing tables are the second argument
}
\examples{
\dontrun{

 # Add tables to the report

del_report <- del_report \%>\%
 add_listing_to_report(
   listing_table(
     pk_delimp_mod  \%>\%
       filter(DELFN == 5, DVID == 2) \%>\%
       arrange(STUDYID, USUBJID, DTTM) \%>\%
       select(ROW,USUBJID, VISIT, PCTPT, PCTESTCD, PCSTAT),
     "Compound A (ng/mL), Sample Not Done / Not Recorded"
   ) \%>\%
     define(ROW, label = " ", width = 0.15, id_var = TRUE, align = "left")
 )  \%>\%
 add_listing_to_report(
   listing_table(
     pk_delimp_mod \%>\%
       filter(DELFN == 5, DVID == 3) \%>\%
       arrange(STUDYID, USUBJID, DTTM) \%>\%
       select(ROW,USUBJID, VISIT, PCTPT, PCTESTCD, PCSTAT),
     "Compound B (ng/mL), Sample Not Done / Not Recorded"
   ))


}
}
