% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mootate.R
\name{mootate}
\alias{mootate}
\title{Mutate and restore variable labels for sweet_read and other labelled objects}
\usage{
mootate(.data, ...)
}
\arguments{
\item{.data}{A data frame (ideally of class "sweet_read")}

\item{...}{Mutations (as in dplyr::mutate)}
}
\value{
A data frame with labels restored.
}
\description{
Works like dplyr::mutate(), but if the input object has class "sweet_read",
restores variable labels from the "var_labels" attribute after mutation using purrr::walk.
Otherwise, uses sjlabelled::copy_labels() to restore labels from the original object.
}
