% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourpaths.R
\name{most_recent}
\alias{most_recent}
\title{most recent}
\usage{
most_recent(path = NULL, pattern = NULL, newest_only = TRUE)
}
\arguments{
\item{pattern}{one or more regex patters e.g. c("dm","lb")}
}
\description{
most recent
}
