% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_disposition.R
\name{pool_disposition}
\alias{pool_disposition}
\title{Make a pooled disposition table}
\usage{
pool_disposition(disp_table)
}
\arguments{
\item{disp_table}{an object of class 'disposition_table'}
}
\value{
A pooled disposition table.
}
\description{
Make a pooled disposition table
}
\examples{
pool_disposition(disp_table = my_disposition_table)
}
