% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_contents.R
\name{print.sweet_contents_tbl}
\alias{print.sweet_contents_tbl}
\title{print.sweet_contents}
\usage{
\method{print}{sweet_contents_tbl}(x, n = 200, ...)
}
\arguments{
\item{x}{object of class sweet_contents_tbl}

\item{n}{maximum number of rows to print}

\item{...}{additional arguments passed to print}
}
\description{
print.sweet_contents
}
