% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourpaths.R
\name{sourpaths}
\alias{sourpaths}
\title{sour paths}
\usage{
sourpaths(full.names = TRUE, path = NULL, ...)
}
\arguments{
\item{full.names}{Default is TRUE}

\item{path}{The directory path to pull from. Default is \code{'../data/'}}

\item{...}{additional arguments passed to list.files()}
}
\value{
a html DT::datatable() with a 'Copy' button
}
\description{
a table of file paths ready to be pasted into an email
}
\examples{
\dontrun{
sourpaths(path="data/",pattern="pk")
}
}
