% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourpaths.R
\name{sweet_asmbdat}
\alias{sweet_asmbdat}
\title{sweet asmbdat}
\usage{
sweet_asmbdat()
}
\value{
TRUE or FALSE
}
\description{
Checks if you're working in asmbdat directory
}
\examples{
\dontrun{
sweet_asmbdat()
}
}
