% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_contents.R
\name{sweet_contents}
\alias{sweet_contents}
\title{Description of all variables}
\usage{
sweet_contents(data, missing_where = NULL)
}
\arguments{
\item{data}{data to observe}

\item{missing_where}{combinations of any variable of choice}
}
\value{
A data frame with name, label (if it exists), n_missing, pct_missing if there are any missings and class.
}
\description{
Description of all variables
}
\examples{
\dontrun{

  broken_pk \%>\%
  sweet_contents()

  broken_pk \%>\%
  sweet_contents(missing_where = DVID)

  broken_pk \%>\%
  sweet_contents(missing_where = VISIT)
}
}
