% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourpaths.R
\name{sweet_sponsor}
\alias{sweet_sponsor}
\title{sweet sponsor}
\usage{
sweet_sponsor()
}
\value{
TRUE or FALSE
}
\description{
Checks if you're working in a client directory
}
\examples{
\dontrun{
sweet_sponsor()
}
}
