% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_listings_report.R
\name{write_listings_report}
\alias{write_listings_report}
\title{Write a listing report}
\usage{
write_listings_report(deletion_report, ...)
}
\arguments{
\item{deletion_report}{A deletion report built using \code{build_report()} and \code{add_listing_to_report()}}

\item{...}{Extra arguments to be passed to \code{reporter::write_report()}}
}
\description{
Output path is set while initializing using \code{build_report()}
This function differs from \code{reporter::write_report()}
by including the N= counts at the end of each listing
}
\examples{
write_listings_report(del_report)



}
