% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourpaths.R
\name{most_recent}
\alias{most_recent}
\title{most recent}
\usage{
most_recent(path = NULL, pattern = NULL, newest_only = TRUE)
}
\arguments{
\item{path}{filepath passed to
\code{path} argument in fs::dir_ls(). Defaut is /dataorig}

\item{pattern}{one or more regex patters e.g. c("dm","lb") to be passed to
\code{regexp} argument in fs::dir_ls()}

\item{newest_only}{return all paths or just the newest default is \code{TRUE}}
}
\value{
data frame
}
\description{
most recent
}
\examples{
\dontrun{
most_recent()
}
}
