% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_contents.R
\name{print.sweet_contents_tbl}
\alias{print.sweet_contents_tbl}
\title{print.sweet_contents}
\usage{
\method{print}{sweet_contents_tbl}(x, ...)
}
\arguments{
\item{x}{object of class sweet_contents_tbl}

\item{...}{print arguments}
}
\description{
print.sweet_contents
}
