% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_disposition.R
\name{sweet_disposition}
\alias{sweet_disposition}
\title{Make a disposition table}
\usage{
sweet_disposition(
  data,
  subjid,
  group_vars,
  cnt_n_keeps = NULL,
  init_desc = "Concentration Records Received"
)
}
\arguments{
\item{data}{a data frame that must include deletion flags called DELFN.}

\item{subjid}{Subject Identifier such as USUBJID, or ID.}

\item{group_vars}{grouping variables useful in disposition tables such as STUDY or STUDYID. If none are provided a default group="1" will be set}

\item{cnt_n_keeps}{Deletion Flags DELFN that should be included in disposition table but do NOT remove subjects or rows.}

\item{init_desc}{Initial description of records received default is "Concentration Records Received"}
}
\value{
A disposition table.
}
\description{
Count the number of subjects and samples impacted by a "deletion flag" DELFN
}
\examples{

broken_pk \%>\%
  dplyr::filter(EVID == 0) \%>\%
  sweet_disposition(subjid = USUBJID, group_vars = STUDYID)


}
